OAK-11731: move MBean interface into oak-api
