+++ oak-api/src/main/java/org/apache/jackrabbit/oak/api/jmx/package-info.java
@@ -14,8 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@Version("4.16.0")
+@Version("4.17.0")
 package org.apache.jackrabbit.oak.api.jmx;
 
 import org.osgi.annotation.versioning.Version;
-
