+++ oak-core/pom.xml
@@ -55,7 +55,6 @@
               org.apache.jackrabbit.oak.plugins.identifier,
               org.apache.jackrabbit.oak.plugins.index,
               org.apache.jackrabbit.oak.plugins.index.aggregate,
-              org.apache.jackrabbit.oak.plugins.index.counter.jmx,
               org.apache.jackrabbit.oak.plugins.index.cursor,
               org.apache.jackrabbit.oak.plugins.index.fulltext,
               org.apache.jackrabbit.oak.plugins.index.importer,
